generic proxy: added support for custom read filter
