+++ contrib/generic_proxy/filters/network/test/BUILD
@@ -64,7 +64,11 @@ envoy_cc_test(
     deps = [
         ":fake_codec_lib",
         "//contrib/generic_proxy/filters/network/source:config",
+        "//contrib/generic_proxy/filters/network/test/mocks:codec_mocks",
+        "//contrib/generic_proxy/filters/network/test/mocks:filter_mocks",
+        "//contrib/generic_proxy/filters/network/test/mocks:route_mocks",
         "//source/common/buffer:buffer_lib",
+        "//test/mocks/network:network_mocks",
         "//test/mocks/server:factory_context_mocks",
         "//test/test_common:registry_lib",
         "//test/test_common:utility_lib",
