+++ contrib/generic_proxy/filters/network/test/fake_codec.h
@@ -257,14 +257,17 @@ public:
 
 class FakeStreamCodecFactoryConfig : public CodecFactoryConfig {
 public:
+  // CodecFactoryConfig
   CodecFactoryPtr createFactory(const Protobuf::Message& config,
                                 Envoy::Server::Configuration::FactoryContext& context) override;
+  FilterFactoryPtr filterFactory(const Protobuf::Message&,
+                                 Server::Configuration::FactoryContext&) override {
+    return nullptr;
+  }
   ProtobufTypes::MessagePtr createEmptyConfigProto() override {
     return std::make_unique<ProtobufWkt::Struct>();
   }
-
   std::set<std::string> configTypes() override { return {"envoy.generic_proxy.codecs.fake.type"}; }
-
   std::string name() const override { return "envoy.generic_proxy.codecs.fake"; }
 };
